Created a separate python script for validation functions called validation.py
